+++ src/components/layout.tsx
@@ -3,16 +3,20 @@ import * as React from "react";
 import { Header } from "./header/header";
 import { Footer } from "./footer/footer";
 
-import { section } from "./layout.css";
+import { Section } from "./section/section";
+import { content } from "./styles.css";
 
 export function Layout(props: React.PropsWithChildren): JSX.Element {
   return (
     <>
-      <Header className={section} />
-      <div className={section}>
-        { props.children }
-      </div>
-      <Footer className={section} />
+      <Section>
+        <Header className={content} />
+      </Section>
+      {props.children}
+      <Section>
+        <Footer className={content} />
+      </Section>
     </>
   );
 }
+
