+++ src/components/section/section.css.ts
@@ -1,0 +1,11 @@
+import { style } from "@vanilla-extract/css";
+import { sprinkles } from "../sprinkles.css";
+
+export const section = style([
+  sprinkles({
+    paddingX: {
+      sm: "large",
+      md: "extraLarge",
+    },
+  }),
+]);
