+++ src/components/section/section.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import { section } from "./section.css";
+
+export function Section(props: React.PropsWithChildren): JSX.Element {
+  return <section className={section}>{props.children}</section>;
+}
